requirements.txt
fastapi==0.111.0
openai==1.30.1
numpy
python-dotenv==1.0.1
pydantic==2.7.1
